JTAF is now version 0.1.2 - Major improvements to the way providers are created.
Provider code is now entirely copied to the working directory that is referenced in the config file. That means all of the .go code is in one place without copying indivial files.

- Commits are handled on terraform destroy

There is now a new commit resource called `destroycommit` which has inverse logic; it commits on a Terraform delete instead of create. This means that when a Terraform module makes use of it, the commit is automatically applied when the resources are destroyed.

- There is now a full example including terraform module best practice sample in the samples directory.

This is our best practice approach.

- The Wiki on the GitHub page is updated. Once this PR is merged, I'll finish the build example steps.

- The go.mod is magically populated into the terraform_providers working directory identified in the config file.

- Testing now with Terraform 0.15.5 and the module 'depends_on' key, which gives us great flexibility.