updated the `terrafrom_providers` file copy functionality to include copying go files in subdirectories well preserving the pathing during the stage where they are copied to the `providerDir`
moved the `config.go` from a file that is copying during processing to a file that is generated during the processing stage, this is so we can properly import subdirectories from the `terraform_providers` directory for example the new `netconf` folder

Added in the `netconf` folder, which contains a nearly identical mirror of the `junos_helper` function from `go-netconf`. With the one exception being that the type has been moved to an interface so the possibility of expanding/altering the `netconf` portion is more accessible, this changed
also required a change to the generated `provider.go` file to reference the new `netconf` helper.